restructure quant ops
